🌏 Geolocation automation :o
Hey, I haven't tested this yet because I made it directly in the browser. However, it should theoretically work, basically what it does is gets the geolocation of a user, then makes a request to a geolocation API to get the country code of the user's longitude and latitude, and from there automatically sets the country and data accordingly.